unit tests from utils.rs moved to tests/unit/utils.rs
